style: delete unnecessary code, format the necessary code and move it to the root
