+++ README.md
@@ -1,2 +1,3 @@
 # todo
 Todo List app
+![Скриншот](https://github.com/Anna-er/todo/blob/main/for-readme/image.jpg)
